Fix pointwise fallback for addcdiv, addcmul.
